+++ packages/shared/src/types/index.ts
@@ -1,1 +1,1 @@
-export type _unstable_mock_type = any;
+export type { Simplify } from './utils';